refactor: change component path
