+++ src/components/loading/index.js
@@ -1,0 +1,11 @@
+import { SLoading, LoadingBox } from './LoadingElements'
+
+const Loading = () => {
+  return (
+    <LoadingBox>
+      <SLoading />
+    </LoadingBox>
+  )
+}
+
+export default Loading
